# swagger.json
{
	"components": {
...
		"schemas": {
...
			"IConversionResult": {
				"properties": {
					"conversionId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"resultFile": {
						"type": "string",
						"format": "byte"
					}
				},
				"required": [
					"conversionId",
					"name",
					"path",
					"resultFile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IConversionStatusResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/IConversionResult"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "converted",
					"result": {
						"conversionId": "55309c37-aed3-4ee9-a143-f1e305333189",
						"name": "testFile",
						"path": "./out/55309c37-aed3-4ee9-a143-f1e305333189.pdf",
						"resultFile": {
							"type": "Buffer",
							"data": [
								67,
								111,
								110,
								118,
								101,
								114,
								116,
								101,
								100,
								32,
								118,
								105,
								97,
								32,
								116,
								101,
								97,
								109,
								112,
								97,
								114,
								97,
								108,
								108,
								97,
								120,
								47,
								117,
								110,
								111,
								99,
								111,
								110,
								118,
								45,
								119,
								101,
								98,
								115,
								101,
								114,
								118,
								105,
								99,
								101,
								10
							]
						}
					}
				}
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "unoconv-webservice",
		"version": "0.1.0",
		"description": "REST api to convert files to different formats using unoconv",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Sven Patrick Meier"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/conversion/{fileId}": {
			"get": {
				"operationId": "GetConvertedFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConversionStatusResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}